requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
idna==3.8
mysql-connector-python==9.0.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
